calculating all relative stats and added async, changed model types to primatives

+++ src/main/java/com/finalfantasy/football/stats/WeekData.java
@@ -3,6 +3,6 @@ package com.finalfantasy.football.stats;
 public class WeekData {
 
   public String statType;
-  public Integer season;
-  public Integer week;
+  public Short season;
+  public Short week;
 }
